implement backoff for reconnect
The current reconnect feature doesn't support backoff, so it's not
really useful since spamming reconnects is not good behaviour.
This introduces a backoff feature, taking inspiration from
https://hex.pm/packages/connection.

This was meant to be a small change, however with the existing
async and handle_initial_conn_failure options, the solution turned out
to be incoherent and also uncovered existing bugs with reconnect during
init that could indefinitely block the supervisor startup process.

As a result, this change makes breaking changes to the api.
In particular, async init is now the default and only behaviour, ie.
a network connection is only made after finalizing the special process
init.